\name{HyRiM-package}
\alias{HyRiM-package}
\alias{HyRiM}
\docType{package}
\title{
\packageTitle{HyRiM}
}
\description{
\packageDescription{HyRiM}
}
%\details{
%
%The DESCRIPTION file:
%\packageDESCRIPTION{HyRiM}
%\packageIndices{HyRiM}
%This package is a prototype/demonstrator implementation of risk management using multiobjective game-theory over abstract spaces of probability distributions. The method is based on results reported in the references below (and references therein).
%}
\author{
\packageAuthor{HyRiM}

Maintainer: \packageMaintainer{HyRiM}
}
\references{
S. Rass, S. Koenig, S. Schauer: Uncertainty in Games: Using Probability-Distributions as Payoffs. in MHR Khouzani et al. (Eds.) GameSec 2015, Springer LNCS 9406, pp. 346-357, DOI: 10.1007/978-3-319-25594-1_20.

S. Rass. On Game-Theoretic Risk Management (Part One). Towards a Theory of
Games with Payoffs that are Probability-Distributions. ArXiv e-prints, June 2015.
http://arxiv.org/abs/1506.07368.

S. Rass. On Game-Theoretic Risk Management (Part Two). Algorithms Algorithms to Compute Nash-Equilibria in Games with Distributions as
Payoffs, ArXiv e-prints, arXiv:1511.08591, 2015.
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\concept{multi-objective game theory}
\concept{stochastic games}
\concept{risk management}
\concept{stochastic order}
\concept{equilibrium}
\concept{security strategy}
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
